+++ previous-versions/assess/RELEASES.md
@@ -391,3 +391,13 @@ Based on:
 - [go v0.26.1] previous-versions/assess
 ### Releases
 - [Go v0.26.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.26.1 - previous-versions/assess
+
+## 2023-09-29 00:12:56
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.27.0] previous-versions/assess
+### Releases
+- [Go v0.27.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.27.0 - previous-versions/assess